Refactored the QueryTypeContentViewMapper
QueryTypeContentViewMapper is now an interface, in order to document both the input and output expectations.
The existing implementation is renamed to QueryParameterContentViewQueryTypeMapper.
