+++ eZ/Publish/Core/QueryType/ContentViewQueryTypeMapper.php
@@ -8,77 +8,15 @@
 namespace eZ\Publish\Core\QueryType;
 
 use eZ\Publish\Core\MVC\Symfony\View\ContentView;
-use InvalidArgumentException;
-use Symfony\Component\ExpressionLanguage\ExpressionLanguage;
 
-class ContentViewQueryTypeMapper implements QueryTypeMapper
+/**
+ * Maps a ContentView to a QueryType.
+ */
+interface ContentViewQueryTypeMapper
 {
     /**
-     * @var QueryTypeRegistry
+     * @param \eZ\Publish\Core\MVC\Symfony\View\ContentView $contentView
+     * @return \eZ\Publish\Core\QueryType\QueryType
      */
-    private $queryTypeRegistry;
-
-    public function __construct(QueryTypeRegistry $queryTypeRegistry)
-    {
-        $this->queryTypeRegistry = $queryTypeRegistry;
-    }
-
-    public function map($contentView)
-    {
-        if (!$contentView instanceof ContentView) {
-            throw new InvalidArgumentException('ContentView expected');
-        }
-
-        if (!$contentView->hasParameter('query')) {
-            throw new InvalidArgumentException('query', "Missing required 'query' view parameter");
-        }
-
-        $queryOptions = $contentView->getParameter('query');
-        $queryType = $this->queryTypeRegistry->getQueryType($queryOptions['query_type']);
-
-        return $queryType->getQuery($this->extractParameters($contentView));
-    }
-
-    /**
-     * @param ContentView $contentView
-     *
-     * @return array
-     */
-    public function extractParameters(ContentView $contentView)
-    {
-        $queryParameters = [];
-
-        $queryOptions = $contentView->getParameter('query');
-        if (isset($queryOptions['parameters'])) {
-            foreach ($queryOptions['parameters'] as $name => $value) {
-                $queryParameters[$name] = $this->evaluateExpression($contentView, $value);
-            }
-        }
-
-        return $queryParameters;
-    }
-
-    /**
-     * @param ContentView $contentView
-     * @param string $queryParameterValue
-     *
-     * @return mixed
-     */
-    public function evaluateExpression(ContentView $contentView, $queryParameterValue)
-    {
-        if (substr($queryParameterValue, 0, 2) === '@=') {
-            $language = new ExpressionLanguage();
-
-            return $language->evaluate(
-                substr($queryParameterValue, 2),
-                [
-                    'view' => $contentView,
-                    'location' => $contentView->getLocation(),
-                    'content' => $contentView->getContent(),
-                ]
-            );
-        } else {
-            return $queryParameterValue;
-        }
-    }
+    public function map(ContentView $contentView);
 }
