--- eZ/Publish/Core/QueryType/QueryTypeMapper.php
@@ -1,13 +1,0 @@
-<?php
-/**
- * This file is part of the ezplatform package.
- *
- * @copyright Copyright (C) eZ Systems AS. All rights reserved.
- * @license For full copyright and license information view LICENSE file distributed with this source code.
- */
-namespace eZ\Publish\Core\QueryType;
-
-interface QueryTypeMapper
-{
-    public function map($input);
-}
